+++ examples/rp2040/examples/now.rs
@@ -1,7 +1,8 @@
-//! Reads the current timestamp and then toggles LED state on  0.5Hz frequency. If the read fails
-//! then program crashes.
+//! Demonstrates RTC functionality by reading the current datetime as a timestamp and then toggling
+//! the LED state on a 0.5Hz frequency.
 //!
-//! Adjust `led_pin` for your board.
+//! See datasheet section 8.6
+
 #![no_std]
 #![no_main]
 
@@ -9,14 +10,14 @@ extern crate nxp_pcf8523;
 
 use embedded_hal::delay::DelayNs;
 use embedded_hal::digital::StatefulOutputPin;
+use nxp_pcf8523::Pcf8523;
 use nxp_pcf8523::datetime::Pcf8523DateTime;
 use nxp_pcf8523::typedefs::Pcf8523T;
-use nxp_pcf8523::Pcf8523;
 use panic_halt as _;
 use rp2040_hal::clocks::init_clocks_and_plls;
 use rp2040_hal::fugit::RateExtU32;
 use rp2040_hal::gpio::{FunctionI2C, FunctionSioOutput, Pin, Pins, PullDown, PullUp};
-use rp2040_hal::{pac, Sio, Timer, Watchdog, I2C};
+use rp2040_hal::{I2C, Sio, Timer, Watchdog, pac};
 
 /// The linker will place this boot block at the start of our program image. We
 /// need this to help the ROM bootloader get our code up and running.